# docs/swagger.yaml
basePath: /
definitions:
  github_com_desen_track-me_internal_dto.APIResponse:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_desen_track-me_internal_dto.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 72
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  github_com_desen_track-me_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_desen_track-me_internal_dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_desen_track-me_internal_dto.UserResponse'
    type: object
...
  github_com_desen_track-me_internal_dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_desen_track-me_internal_dto.RegisterRequest:
    properties:
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - display_name
    - email
    - password
    type: object
...
  github_com_desen_track-me_internal_dto.UserResponse:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      status_message:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Backend REST & WebSocket API for lightweight, privacy-focused mutual
    location sharing.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mutual Location Sharing API
  version: "1.0"
paths:
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Update password for the authenticated user and invalidate all other
        sessions
      parameters:
      - description: Change Password Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Validation or mismatch error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized or incorrect old password
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Auth
  /api/v1/auth/delete:
    delete:
      consumes:
      - application/json
      description: Permanently delete the authenticated user account and all session
        data
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return token pair
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_desen_track-me_internal_dto.LoginResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate current user session and refresh token
      parameters:
      - description: Refresh Token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access and refresh token pair using a valid refresh token
      parameters:
      - description: Refresh Token Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_desen_track-me_internal_dto.LoginResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Invalid or expired session
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      summary: Refresh token pair
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new account with email, password, and display name
      parameters:
      - description: Register Request Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_desen_track-me_internal_dto.UserResponse'
              type: object
        "400":
          description: Validation or duplicate email error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token (e.g. Bearer <token>)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"